requirements.txt
# Dev Dependencies
pycodestyle==2.4.0
pytest==3.5.1
python-Levenshtein==0.12.0
twine==1.11.0

# Dependencies
numpy==1.14.3
